Move Ember.Component to components/ directory
