+++ packages/ember-htmlbars/tests/compat/view_helper_test.js
@@ -1,4 +1,4 @@
-import EmberComponent from 'ember-views/views/component';
+import EmberComponent from 'ember-views/components/component';
 import EmberView from 'ember-views/views/view';
 import EmberSelectView from 'ember-views/views/select';
 import { runAppend, runDestroy } from 'ember-runtime/tests/utils';
