+++ packages/ember-htmlbars/tests/compat/view_keyword_test.js
@@ -1,4 +1,4 @@
-import EmberComponent from 'ember-views/views/component';
+import EmberComponent from 'ember-views/components/component';
 import { runAppend, runDestroy } from 'ember-runtime/tests/utils';
 import compile from 'ember-template-compiler/system/compile';
