+++ packages/ember-htmlbars/tests/helpers/component_test.js
@@ -7,7 +7,7 @@ import Registry from 'container/registry';
 import { runAppend, runDestroy } from 'ember-runtime/tests/utils';
 import ComponentLookup from 'ember-views/component_lookup';
 import EmberView from 'ember-views/views/view';
-import Component from 'ember-views/views/component';
+import Component from 'ember-views/components/component';
 import compile from 'ember-template-compiler/system/compile';
 import computed from 'ember-metal/computed';
