+++ packages/ember-htmlbars/tests/helpers/custom_helper_test.js
@@ -1,5 +1,5 @@
 import Ember from 'ember-metal/core';
-import Component from 'ember-views/views/component';
+import Component from 'ember-views/components/component';
 import Helper, { helper as makeHelper } from 'ember-htmlbars/helper';
 import compile from 'ember-template-compiler/system/compile';
 import { runAppend, runDestroy } from 'ember-runtime/tests/utils';
