+++ packages/ember-htmlbars/tests/integration/helper-lookup-test.js
@@ -2,7 +2,7 @@ import isEnabled from 'ember-metal/features';
 import Registry from 'container/registry';
 import compile from 'ember-template-compiler/system/compile';
 import ComponentLookup from 'ember-views/component_lookup';
-import Component from 'ember-views/views/component';
+import Component from 'ember-views/components/component';
 import { helper } from 'ember-htmlbars/helper';
 import { runAppend, runDestroy } from 'ember-runtime/tests/utils';
