+++ packages/ember-htmlbars/tests/integration/mutable_binding_test.js
@@ -4,7 +4,7 @@ import Registry from 'container/registry';
 //import jQuery from "ember-views/system/jquery";
 import compile from 'ember-template-compiler/system/compile';
 import ComponentLookup from 'ember-views/component_lookup';
-import Component from 'ember-views/views/component';
+import Component from 'ember-views/components/component';
 import { runAppend, runDestroy } from 'ember-runtime/tests/utils';
 import run from 'ember-metal/run_loop';
 import { computed } from 'ember-metal/computed';
