+++ packages/ember-htmlbars/tests/integration/will-destroy-element-hook-test.js
@@ -1,5 +1,5 @@
 import run from 'ember-metal/run_loop';
-import Component from 'ember-views/views/component';
+import Component from 'ember-views/components/component';
 import compile from 'ember-template-compiler/system/compile';
 import { runAppend, runDestroy } from 'ember-runtime/tests/utils';
