+++ packages/ember-routing-views/lib/views/link.js
@@ -11,7 +11,7 @@ import { set } from 'ember-metal/property_set';
 import { computed } from 'ember-metal/computed';
 import { deprecatingAlias } from 'ember-metal/computed_macros';
 import { isSimpleClick } from 'ember-views/system/utils';
-import EmberComponent from 'ember-views/views/component';
+import EmberComponent from 'ember-views/components/component';
 import inject from 'ember-runtime/inject';
 import 'ember-runtime/system/service'; // creates inject.service
 import ControllerMixin from 'ember-runtime/mixins/controller';
