+++ packages/ember-views/lib/components/component.js
