+++ packages/ember-views/lib/views/checkbox.js
@@ -1,6 +1,6 @@
 import { get } from 'ember-metal/property_get';
 import { set } from 'ember-metal/property_set';
-import EmberComponent from 'ember-views/views/component';
+import EmberComponent from 'ember-views/components/component';
 
 /**
 @module ember
