+++ packages/ember-views/tests/views/component_test.js
@@ -7,7 +7,7 @@ import inject from 'ember-runtime/inject';
 import { get } from 'ember-metal/property_get';
 
 import EmberView from 'ember-views/views/view';
-import Component from 'ember-views/views/component';
+import Component from 'ember-views/components/component';
 
 import { MUTABLE_CELL } from 'ember-views/compat/attrs-proxy';
